fix: 서비스 레이어 인터페이스 추가
